LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpTransceiver.https-expected.txt


Harness Error (TIMEOUT), message = null

PASS checkAddTransceiverNoTrack 
PASS checkAddTransceiverWithTrack 
PASS checkAddTransceiverWithAddTrack 
PASS checkAddTransceiverWithDirection 
PASS checkAddTransceiverWithSetRemoteOfferSending 
PASS checkAddTransceiverWithSetRemoteOfferNoSend 
PASS checkAddTransceiverBadKind 
FAIL checkNoMidOffer promise_test: Unhandled rejection with value: object "InvalidAccessError: Failed to set remote offer sdp: The BUNDLE group contains MID:0 matching no m= section."
PASS checkSetDirection 
PASS checkCurrentDirection 
PASS checkSendrecvWithNoSendTrack 
PASS checkSendrecvWithTracklessStream 
PASS checkAddTransceiverNoTrackDoesntPair 
PASS checkAddTransceiverWithTrackDoesntPair 
PASS checkAddTransceiverThenReplaceTrackDoesntPair 
FAIL checkAddTransceiverThenAddTrackPairs assert_equals: expected "[{sender:{track:{}}}]" but got "[{sender:{track:{}}},{}]"
PASS checkAddTrackPairs 
PASS checkReplaceTrackNullDoesntPreventPairing 
PASS checkRemoveAndReadd 
PASS checkAddTrackExistingTransceiverThenRemove 
FAIL checkRemoveTrackNegotiation promise_test: Unhandled rejection with value: object "InvalidStateError: Description type incompatible with current signaling state"
TIMEOUT checkMute Test timed out
NOTRUN checkStop 
NOTRUN checkStopAfterCreateOffer 
NOTRUN checkStopAfterSetLocalOffer 
NOTRUN checkStopAfterSetRemoteOffer 
NOTRUN checkStopAfterCreateAnswer 
NOTRUN checkStopAfterSetLocalAnswer 
NOTRUN checkStopAfterClose 
NOTRUN checkLocalRollback 
NOTRUN checkRollbackAndSetRemoteOfferWithDifferentType 
NOTRUN checkRemoteRollback 
NOTRUN checkMsectionReuse 
